Add CSS to foundations exercise class id selectors
This commit completes exercise 2: class-id-selectors of TOP CSS
exercises foundations.
Before this commit the style.css file was empty and the contents of the
 webpage rendered by the browser when reading the index.html file were
not styled.
Classes and IDs were added to the html elements and font and color
styles were added. Now the rendered html file is styled according to
the desired-outcome.jpg