Change background-color value from HSL to HSLA
This commit will change the background-color property's value from HSL
to HSLA in the odd class, inside the style.css file.
Before the value was using HSL color system to specify the color value.
This meant that the fourth attribute provided, transparency alpha, was
not having an effect on the color of the background of elements of
class odd.