Complete exercise 04-chaining-selectors
This commit completes exercises 4 from the Odin Project CSS foundations
exercises.
This exercise is about chaining selectors and image size editing using
CSS.
Before the images in the HTML file were not styles using CSS, this
meant that they retained their original image sizes, width and height
measured in pixels.
The top two images were meant to have their sizes changed, matching
their class attribute names in their html img opening tags.
Both image elements were part of the "avatar" class, but the first one
was also part of the "proportioned" class, and the second one was also
part of the "distorted" class.
This commit edits the first 2 images so that the first one is
proportioned and 300 pixels wide, while the second one is distorted and
200 pixels wide and twice as high.
This is the desired outcome.
This is done according to the exercise requirement of using chaining
selectors. So, images with the proportioned class AND the avatar
class are styled in the desired proportioned avatar manner, and images
in the distorted class AND the avatar class are styled in the desired
distorted avatar manner. While images with the proportioned class or the
distorted class attribute, that DO NOT also have the avatar class, are
not styled in this way. For example, the third and fourth images are
elements with the proportioned and distorted class, but they are not
with the avatar class. So they should not, and are not, styled by the
.avatar.proportioned or .avatar.distorted selector.