Complete css foundations ex5-descendant combinator
This commit implements the solution to exercise 5 using a descendant
combinator to combine multiple selectors depending on their
relationship in a CSS rule. In this case the relationship between
selectors is that of descendant, hence the name descendant combinator.
The descendant combinator in the commit used both a type selector and a
class selector. Different solutions were tried using just type
selectors, just class selectors, or a different combination of the two.